requirements.txt
# OpenAI API for embeddings and LLM
openai>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Vector similarity and KNN
scikit-learn>=1.3.0

# Environment management
python-dotenv>=1.0.0

# Optional: Progress bars and utilities
tqdm>=4.65.0

# Visualization (for data analysis)
matplotlib>=3.7.0

